Follow the user's JAX precision instead of forcing x64 on import
A library shouldn't mutate the global jax.config: enabling x64 at import is
process-wide (slows GPU/TPU work, surprises other libraries). Remove the
jax.config.update call from __init__ and replace all 342 hard-coded
dtype=jnp.float64 with dtype=float, which JAX canonicalizes to the active default
(float32 by default, float64 when the user enables x64) with no truncation
warning. Import is now side-effect-free and precision tracks the user's flag.

For magpylib parity, enable x64 before use: jax.config.update('jax_enable_x64', True).
The test suite keeps x64 (conftest). tests/test_precision.py pins the contract
(import doesn't enable x64; float32 default, no warning; float64 opt-in). README,
quickstart, and numerics docs updated with a Precision section. BREAKING: default
precision is now float32.

Thanks to the collaborator who flagged this.

+++ src/magpylib_jax/core/kernels/_raycast.py
@@ -136,8 +136,8 @@ def _mask_inside_trimesh_jax(points: jnp.ndarray, faces: jnp.ndarray) -> jnp.nda
     mz = (points[:, 2] < zmax + eps) & (points[:, 2] > zmin - eps)
     mask_box = mx & my & mz
 
-    start_point_outside = jnp.array([xmin, ymin, zmin], dtype=jnp.float64) - jnp.array(
-        [12.0012345, 5.9923456, 6.9932109], dtype=jnp.float64
+    start_point_outside = jnp.array([xmin, ymin, zmin], dtype=float) - jnp.array(
+        [12.0012345, 5.9923456, 6.9932109], dtype=float
     )
     start_pts = jnp.broadcast_to(start_point_outside, points.shape)
     lines = jnp.stack((start_pts, points), axis=1)
@@ -167,8 +167,8 @@ def _mask_inside_trimesh_jax_masked(
         mz = (points[:, 2] < zmax + eps) & (points[:, 2] > zmin - eps)
         mask_box = mx & my & mz
 
-        start_point_outside = jnp.array([xmin, ymin, zmin], dtype=jnp.float64) - jnp.array(
-            [12.0012345, 5.9923456, 6.9932109], dtype=jnp.float64
+        start_point_outside = jnp.array([xmin, ymin, zmin], dtype=float) - jnp.array(
+            [12.0012345, 5.9923456, 6.9932109], dtype=float
         )
         start_pts = jnp.broadcast_to(start_point_outside, points.shape)
         lines = jnp.stack((start_pts, points), axis=1)
